+++ src/main/java/io/bootique/liquibase/LiquibaseRunner.java
@@ -13,7 +13,6 @@ import liquibase.database.DatabaseFactory;
 import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.DatabaseException;
 import liquibase.exception.LiquibaseException;
-import liquibase.logging.LogFactory;
 import liquibase.resource.ResourceAccessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -32,7 +31,6 @@ public class LiquibaseRunner {
     protected Collection<ResourceFactory> changeLogs;
     protected DataSource dataSource;
     protected Cli cli;
-    private SLFLiquibaseAdapter loggerAdapter;
 
     public LiquibaseRunner(Collection<ResourceFactory> changeLogs,
                            DataSource dataSource,
@@ -40,11 +38,6 @@ public class LiquibaseRunner {
         this.changeLogs = Objects.requireNonNull(changeLogs);
         this.dataSource = Objects.requireNonNull(dataSource);
         this.cli = cli;
-
-        Logger lbLogger = LoggerFactory.getLogger(Liquibase.class);
-        this.loggerAdapter = new SLFLiquibaseAdapter(lbLogger);
-
-        setupLogging();
     }
 
     public <T> T runWithLiquibase(Function<Liquibase, T> op) {
@@ -69,15 +62,6 @@ public class LiquibaseRunner {
         }
     }
 
-    protected void setupLogging() {
-        LogFactory.setInstance(new LogFactory() {
-            @Override
-            public liquibase.logging.Logger getLog(String name) {
-                return loggerAdapter;
-            }
-        });
-    }
-
     protected DatabaseChangeLog createDatabaseChangeLog(Database database, ResourceAccessor resourceAccessor) {
         DatabaseChangeLog changeLog = new DatabaseChangeLog();
         changeLog.setChangeLogParameters(new ChangeLogParameters(database));
@@ -85,7 +69,7 @@ public class LiquibaseRunner {
         changeLogs.forEach(cl -> {
             try {
                 LOGGER.info("Including change log: '{}'", cl.getResourceId());
-                changeLog.include(cl.getResourceId(), false, resourceAccessor, new ContextExpression());
+                changeLog.include(cl.getResourceId(), false, resourceAccessor, new ContextExpression(), true);
             } catch (LiquibaseException e) {
                 throw new RuntimeException("Error configuring Liquibase", e);
             }